fix: 엔티티에 맞게 repository 의 id 변경

+++ src/main/java/art/snail/naillian/backend/domain/nail/repository/NailSetRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 
-public interface NailSetRepository extends ReactiveCrudRepository<NailSet, Long> {
+public interface NailSetRepository extends ReactiveCrudRepository<NailSet, Integer> {
     Flux<NailSet> findAllBy(Pageable page);
 
 }
